+++ Lab_4_V1_2017_09_21/Key.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Lab_4_V1_2017_09_21
+{
+
+    class Key : Squar
+    {
+
+        public Key()
+        {
+            this.MapChar = 'K';
+            this.MapColor = ConsoleColor.Blue;
+            this.TypAvSquar = TyperAvSquarse.Key;
+            this.CanMoveOnSquare = true;
+        }
+
+    }
+}
