+++ Lab_4_V1_2017_09_21/Lab_4_V1_2017_09_21.csproj
@@ -41,9 +41,19 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Highscore.cs" />
+    <Compile Include="Door.cs" />
+    <Compile Include="EmptySquar.cs" />
+    <Compile Include="Exit.cs" />
+    <Compile Include="Key.cs" />
     <Compile Include="Map.cs" />
+    <Compile Include="Monster.cs" />
+    <Compile Include="Player.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Room.cs" />
+    <Compile Include="Squar.cs" />
+    <Compile Include="Walls.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
