+++ Lab_4_V1_2017_09_21/Program.cs
@@ -77,7 +77,6 @@ namespace Lab_4_V1_2017_09_21
             Console.WriteLine("The End");
             Console.WriteLine("The End");
             Console.WriteLine("The End");
-            Console.ReadLine();
         }
 
         static void MoveCurserToDefultePos()
